fix(events): invalidate leaf-level administrative area cache on write
